+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
@@ -6,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.0.1] - 2018-02-20
+## [0.0.2] - 2018-03-10
+
 ### Added
+
+- Get all torrent shows on sync mode
+- Show modal window on loading
+
+## [0.0.1] - 2018-02-20
+
+### Added
+
 - First big commit